+++ YuzSDK/Utils/Models/Main/InvoiceItemModel.swift
@@ -13,7 +13,7 @@ public protocol InvoiceItemProtocol: ModelProtocol {
     var branchName: String? {get set}
     var organizationName: String? {get set}
     var clientName: String? {get set}
-    var totalAmount: Int? {get set}
+    var totalAmount: Float? {get set}
     var invoiceNote: String? {get set}
     var createdDate: String? {get set}
     var isPaid: Bool? {get set}
@@ -28,11 +28,11 @@ public struct InvoiceItemModel: ModelProtocol, InvoiceItemProtocol {
     
     public var invoiceID: Int
     public var operatorName, branchName, organizationName, clientName: String?
-    public var totalAmount: Int?
+    public var totalAmount: Float?
     public var invoiceNote, createdDate: String?
     public var isPaid, isExpired: Bool?
     
-    public init(invoiceID: Int, operatorName: String? = nil, branchName: String? = nil, organizationName: String? = nil, clientName: String? = nil, totalAmount: Int? = nil, invoiceNote: String? = nil, createdDate: String? = nil, isPaid: Bool? = nil, isExpired: Bool? = nil) {
+    public init(invoiceID: Int, operatorName: String? = nil, branchName: String? = nil, organizationName: String? = nil, clientName: String? = nil, totalAmount: Float? = nil, invoiceNote: String? = nil, createdDate: String? = nil, isPaid: Bool? = nil, isExpired: Bool? = nil) {
         self.invoiceID = invoiceID
         self.operatorName = operatorName
         self.branchName = branchName
